move netcdf_io to common
