+++ src/Common/netcdf_io.jl
@@ -11,9 +11,9 @@ end
 function NetCDFIO_Stats(
     nc_filename,
     output_interval,
-    x,
     z_faces,
     z_centers;
+    x = nothing,
     output_profiles = Dict(
         :ρ => "density",
         :T => "temperature",
@@ -29,32 +29,46 @@ function NetCDFIO_Stats(
 )
     FT = Float64
 
+    dim = x === nothing ? 1 : 2
+
     # Remove the NC file if it exists, in case it accidentally wasn't closed
     isfile(nc_filename) && rm(nc_filename; force = true)
 
     NC.Dataset(nc_filename, "c") do ds
         # Set profile dimensions
         profiles_grp = NC.defGroup(ds, "profiles")
-        NC.defDim(profiles_grp, "x", length(x))
         NC.defDim(profiles_grp, "zf", length(z_faces))
         NC.defDim(profiles_grp, "zc", length(z_centers))
         NC.defDim(profiles_grp, "t", Inf)
-        NC.defVar(profiles_grp, "x", x, ("x",))
         NC.defVar(profiles_grp, "zf", z_faces, ("zf",))
         NC.defVar(profiles_grp, "zc", z_centers, ("zc",))
         NC.defVar(profiles_grp, "t", FT, ("t",))
 
+        if dim == 2
+            NC.defDim(profiles_grp, "x", length(x))
+            NC.defVar(profiles_grp, "x", x, ("x",))
+        end
+
         for var in keys(output_profiles)
-            NC.defVar(profiles_grp, output_profiles[var], FT, ("x", "zc", "t"))
+            if dim == 1
+                NC.defVar(profiles_grp, output_profiles[var], FT, ("zc", "t"))
+            elseif dim == 2
+                NC.defVar(profiles_grp, output_profiles[var], FT, ("x", "zc", "t"))
+            else
+                error("Invalid system dimension!!!")
+            end
         end
 
         reference_grp = NC.defGroup(ds, "reference")
-        NC.defDim(reference_grp, "x", length(x))
         NC.defDim(reference_grp, "zf", length(z_faces))
         NC.defDim(reference_grp, "zc", length(z_centers))
-        NC.defVar(reference_grp, "x", x, ("x",))
         NC.defVar(reference_grp, "zf", z_faces, ("zf",))
         NC.defVar(reference_grp, "zc", z_centers, ("zc",))
+
+        if dim == 2
+            NC.defDim(reference_grp, "x", length(x))
+            NC.defVar(reference_grp, "x", x, ("x",))
+        end
 
         timeseries_grp = NC.defGroup(ds, "timeseries")
         NC.defDim(timeseries_grp, "t", Inf)
@@ -64,9 +78,10 @@ function NetCDFIO_Stats(
     return NetCDFIO_Stats(output_interval, nc_filename, output_profiles)
 end
 
-function Kinematic2D_output(aux, t::FT) where {FT <: Real}
+function simulation_output(aux, t::FT) where {FT <: Real}
 
     (; Stats) = aux
+    dim = :domain_width in keys(aux) ? 2 : 1
 
     NC.Dataset(Stats.nc_filename, "a") do ds
 
@@ -83,7 +98,13 @@ function Kinematic2D_output(aux, t::FT) where {FT <: Real}
                     prop = getproperty(field, var)
                     # _parentname = nothing to avoid https://github.com/Alexander-Barth/NCDatasets.jl/issues/135
                     ncdf_var = NC.cfvariable(ds.group["profiles"], Stats.output_profiles[var], _parentname = nothing)
-                    @inbounds ncdf_var[:, :, end] = parent(prop)[:, 1, 1, :]
+                    if dim == 1
+                        @inbounds ncdf_var[:, end] = vec(prop)
+                    elseif dim == 2
+                        @inbounds ncdf_var[:, :, end] = parent(prop)[:, 1, 1, :]
+                    else
+                        error("Invalid system dimension!!!")
+                    end
                 end
             end
         end
@@ -98,8 +119,23 @@ end
 """
    Interface to ODE callbacks for handling output.
 """
+
+function condition_io(u, t, integrator)
+    UnPack.@unpack TS, Stats = integrator.p
+    TS.dt_io += TS.dt
+    io_flag = false
+    if TS.dt_io > Stats.output_interval
+        TS.dt_io = 0
+        io_flag = true
+    end
+    return io_flag || t ≈ 0 || t ≈ TS.t_max
+end
+
+"""
+   Interface to ODE callbacks for handling output.
+"""
 function affect_io!(integrator)
-    Kinematic2D_output(integrator.p, integrator.t)
+    simulation_output(integrator.p, integrator.t)
 
     ODE.u_modified!(integrator, false) # We're legitamately not mutating `u` (the state vector)
 end
