+++ src/K1DModel/K1DModel.jl
@@ -1,9 +1,7 @@
 module K1DModel
 
 import OrdinaryDiffEq as ODE
-import NCDatasets as NC
 import SpecialFunctions as SF
-import UnPack
 import Logging
 import TerminalLoggers
 
@@ -18,8 +16,6 @@ import ..Common as CO
 
 include("parameters.jl")
 include("ode_utils.jl")
-include("netcdf_io.jl")
-include("callbacks.jl")
 include("initial_condition.jl")
 include("tendency.jl")
